experimental export marking date to excel script
It does work but there are some limitations and it's not totally clear
how to work round them (see below for details). I suspect the answer is
going to be a tab with extra cells that have longer option lists in.
Unsure how to handle multiple select though as not really supported in
Excel except via quick hacky solutions or VBA.

TODO:
 * handle multiple select answers
 * handle long options (excel has a 255 char limit for lists)
 * add in all answer boxes etc