Link to council section audit from volunteer_progress when in audit phase
Previously, even when the audit phase was active, clicking a council’s
name on the /volunteer_progress/<userid> page would take you to the
first mark form for that authority+section, rather than the audit form
for that authority+section.

We now alter the destination of the links on /volunteer_progress/<userid>
based on the currently active mark phase.