Fix “Copy from First Mark” button for FOIs
FOI links were stored in the `evidence` field during First Mark, but
CEUK want them transferred to the `public_notes` field during Audit
(so that they are visible to the public).

This commit adds an exception for FOI questions, that redirects the
content to the right fields, when you press the “Copy from First Mark”
button on FOI questions.

Fixes #87.