attempt to reduce duplicate sumissions and hence responses
Given I've not been able to replicate the underlying issue I've
essentially taken a best guess approach to what the problem is based on
the evidence and written a fix that should resolve that. It's not
totally foolproof in that it assumes that the duplicate submissions are
identical but I don't understand how they cannot be.

It doesn't throw up an error, just silently throws away the identical
data and carries on which I think is fine as then end result will be the
same for the user and this way they don't get an error message they
can't do anything about.

Fixes #40