add optional logging of failed login attempts
If LOG_FAILED_LOGINS is True in .env then this will use the
user_login_failed signal to send some basic details about failed logins
if logging is configured. Hopefully helpful to try and work out why
users are unable to log in. Note that this does not include the reason
shown to the user but does say if the account is present, active and has
a marker object.