make national data script more flexible and session compliant
This makes a number of upddates to the national data script most of
which are to allow it to work with sessions, but also to allow a bit
more flexibilty. They are:

* add session support
* move the config out to a json file versus hard coding it
* pass in the data location as an argument
* do not commit the data by default

The script also no longer creates the options as that should have
happened as part of the question import.

There are also various small changes to the type of data that can be
included in the file:

* use GSS codes directly to look up councils
* better skip data options
* mostly uses the score to match options, but can use desc if score_desc
  set in question config

update national data import to respect marking sessions

also add an option to make setting up options optional

make file location and config location command arguments

also change the config to being loaded from a json file instead of hard
coded.

do not add responses by default for national data import

should make testing it a bit less fraught

handle sheets having GSS code directly in them

No need for a council lookup so do not populate it and get the GSS code
directly from the data sheet

allow only importing national data if a column doesn't matches a value

Some sheets have data for, e.g. multiple years in and we only want one
of them so update the skip logic to allow for skipping unless the column
matches versus skipping if the column does not match

allow matching option using score only in national data import

as we're not creating the options from script it makes more sense to
match using the score than the description as then don't need to do a
lot of redundant config