allow authority response_counts to ignore the question list
This is to fix issues where you want response counts from a set of mixed
question group authorities and it's not possible to pass in a question
list that fits all of them so we fall back to using the question group
to match the question list.

However, keep a question list as default for backwards compatability.