add ability to cache all assignment progress for a session
Moves the calculating of progress out to a method in a marking file,
uses that in a management command to store all progress for all types in
a pickle file and if that exists loads that on the front page.

This should significantly speed up the display of the front page for
admin users. For non admin users progress is loaded from the database as
before.